test: split test files in server and browser
